feat: rename services to modules
