+++ src/modules/band/resolvers/bands.resolver.ts
