+++ src/modules/favourites/resolvers/favourites.resolver.ts
